SR_Data/requirements.txt
altair==5.4.1
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
branca==0.7.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
flake8==7.1.1
folium==0.17.0
fonttools==4.53.1
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
investiny==0.7.2
investpy==1.0.8
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
joblib==1.4.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.6
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
multitasking==0.0.11
narwhals==1.6.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.7
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
plotly==5.24.1
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==1.10.18
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.2
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.8.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
streamlit==1.38.0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
widgetsnbextension==4.0.13
wsproto==1.2.0
xyzservices==2024.9.0
yfinance==0.2.46
